requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
tensorflow~=2.16.1
pygame~=2.5.2
pyttsx3~=2.90
soundfile~=0.12.1
mutagen~=1.47.0
pydub~=0.25.1
urllib3~=2.2.1
requests~=2.31.0
python-dotenv~=1.0.1
pillow~=10.2.0
pynput~=1.7.7
python-vlc~=3.0.20123
